pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
  "flask>=3.0.0",
  "python-dotenv>=1.0.1",
  "gunicorn>=23.0.0",
  "psycopg2-binary>=2.9.11",
  "validators>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
